fix: prevent undesired unsubscribing and unreviewing when deleting content, WEB-892

+++ app/models/identification.rb
@@ -126,7 +126,7 @@ class Identification < ApplicationRecord
 
   auto_subscribes :user, to: :observation, if: lambda {| ident, observation |
     ident.user_id != observation.user_id
-  }
+  }, unsubscribe_enabled: :observation_unsubscribe_enabled?
 
   notifies_users :mentioned_users,
     on: :save,
@@ -427,6 +427,8 @@ class Identification < ApplicationRecord
   end
 
   def remove_automated_observation_reviews
+    return if Identification.where( observation_id: observation_id, user_id: user_id ).where.not( id: id ).any?
+
     ObservationReview.where( observation_id: observation_id,
       user_id: user_id, user_added: false ).destroy_all
     true
@@ -581,6 +583,14 @@ class Identification < ApplicationRecord
 
   def taxon_rank
     taxon.try( :rank )
+  end
+
+  def observation_unsubscribe_enabled?
+    return false if Identification.where( observation: observation, user_id: user_id ).where.not( id: id ).any?
+    return false if Comment.where( parent: observation, user_id: user_id ).any?
+    return false if ObservationFieldValue.where( observation: observation, user_id: user_id ).any?
+
+    true
   end
 
   # Static ##################################################################
